+++ src/BaseClasses/Map.java
@@ -1,0 +1,4 @@
+package BaseClasses;
+
+public class Map {
+}
