+++ src/BaseClasses/Tile.java
@@ -1,3 +1,5 @@
+package BaseClasses;
+
 import Entities.*;
 import java.util.*;
