+++ src/Entities/Entity.java
@@ -1,13 +1,28 @@
 package Entities;
 
-/* will store
-a refrence to a tile
-and a ref to a region
-*/
-public class Entity {
-    public Entity(){
+import BaseClasses.Position;
+import BaseClasses.Region;
 
+public class Entity {
+
+    private Position position;
+
+    public Position getPosition(){
+        return position;
+    }
+
+    public void setPosition(Position position){
+        //Important that the address doesn't change but the values do
+        this.position.set(position);
     }
 
 
+    //Constructor for using reference (equipped/held items)
+    public Entity(Position position){   
+        this.position = position;
+    }
+
+    public Entity(Region region, int x, int y){
+        this.position = new Position(region, x, y);
+    }
 }
