+++ src/Entities/Mobiles/Player.java
@@ -1,7 +1,5 @@
 package Entities.Mobiles;
 
-import Entities.Entity;
-
 public class Player extends Mobile {
 
 }
