--- src/Map.java
@@ -1,2 +1,0 @@
-public class Map {
-}
